Original source from ispell-norsk-2.0.tar.gz at http://folk.uio.no/runekl/dictionary.html (which is no longer maintained.)
I have a patch that updates this for MySpell/OpenOffice and aspell support,
and also cleans up the build. The idea is to check in the patch (possibly
copying its entire hostory from CVS server at procaptura.com) along with
these sources, apply the patch to get an up-to-date source on the trunc, then
throw away the patch and start modifying the sources directly. There will be
a separate branch for runekl's sources, although I think we can be pretty
sure that there will be no further changes there. (I've in practice
taken over the project.)
 
Note that the original setup assumes these files are unpackacked in sub
directory "languages/norsk" of the ispell source code distribution (or that
the .tar.gz is extracted in "languages" - the archive includes directory
level "norsk".)