All my earlier updates patched in. Command was
 patch -p3 < patches/ispell-norsk.myspell.patch

with top of CVS tree as current dir.

For those who didn't know: I have so far keept all my changes in a single
patch file, checked in as patches/ispell-norsk.myspell.patch on this tree.
I'll now be storing the actual source files in CVS instead, and will apply all
changes directly. This means there are many "versions" between this revision
and the original check-in - please consult the above mentioned patch file for
details.