norsk.words.sq replaced by plain-text variant norsk.words, so that we don't have to rely on special "unsq" tool, which isn't even included in the latest ispell.
That's one large file, but hopefully it is CVS-friendlier than .sq version.

I also considered checking in the variant where list is split up into multiple
files according to cathegory, but decided not to, as the split-up rule contains
comment removal/sort, which we would need commands for in any case.

+ "clean" cleaned up (!) a bit.